+++ kinto_http/utils.py
@@ -1,5 +1,8 @@
+import functools
+import json
 import re
 import unicodedata
+from datetime import date, datetime
 from unidecode import unidecode
 
 
@@ -44,3 +47,13 @@ def chunks(l, n):
             yield l[i:i+n]
     else:
         yield l
+
+
+def json_iso_datetime(obj):
+    """JSON serializer for objects not serializable by default json code"""
+    if isinstance(obj, (datetime, date)):
+        return obj.isoformat()
+    raise TypeError("Type %s is not serializable" % type(obj))
+
+
+json_dumps = functools.partial(json.dumps, default=json_iso_datetime)